// src/types/entities/ConsultExampleCommentEntity.ts
export type ConsultExampleCommentEntity = {
  consult_example_comment_id: string;
  example_message_id: number;
  body: string;
  is_anonymous: boolean;
  doctor_last_name: string;
  doctor_first_name: string;
  speciality: string;
  qualified_year: number;
  like_count: number;
  comment_count: number;
  is_liked: boolean;
  created_date: string;
};